add common module, shift RequestService, LoadingSpinner into it remove welcome and referral page
