+++ beeline/common/main.js
@@ -1,0 +1,6 @@
+let common = angular.module("common", [])
+
+require("./RequestService")
+require("./LoadingSpinner")
+
+module.exports = common
