--- beeline/templates/share-referral-modal.html
@@ -1,23 +1,0 @@
-<ion-modal-view>
-  <ion-header-bar class="bar-royal bar bar-header">
-    <h1 class="title">Invite friends</h1>
-    <button class="button button-clear icon ion-ios-close-outline" ng-click="shareReferralModal.hide()">
-    </button>
-  </ion-header-bar>
-  <ion-content id="share-rides">
-    <div class="item item-text-wrap">
-      <div class="referral-code">
-        {{user.referralCode.code}}
-      </div>
-      <p>Share this to your friend so that they receive $10 off Beeline rides.  You'll automatically get $10 ride credits, once they claim their first ride.</p>
-      <div class="share-box">
-        <textarea rows="4" id="shareMsg" ng-model="shareMsg"></textarea>
-
-        <div class="text-right">
-          <button ng-if="!hasCordova" class="button button-outline button-royal small-button" ngclipboard data-clipboard-text="{{ shareMsg }}">Copy</button>
-          <button ng-if="hasCordova" class="button button-outline button-royal small-button" ng-click="cordovaShare()">Share</button>
-        </div>
-      </div>
-    </div>
-  </ion-content>
-</ion-modal-view>
